+++ R/id.R
@@ -1,8 +1,4 @@
 id <- function(.variables, drop = FALSE) {
-  # Drop all zero length inputs
-  lengths <- vapply(.variables, length, integer(1))
-  .variables <- .variables[lengths != 0]
-
   if (length(.variables) == 0) {
     n <- nrow(.variables) %||% 0L
     return(structure(seq_len(n), n = n))
@@ -43,12 +39,14 @@ id <- function(.variables, drop = FALSE) {
 }
 
 id_var <- function(x, drop = FALSE) {
-  if (length(x) == 0) return(structure(integer(), n = 0L))
   if (!is.null(attr(x, "n")) && !drop) return(x)
 
   if (is.factor(x) && !drop) {
     id <- as.integer(addNA(x, ifany = TRUE))
     n <- length(levels(x))
+  } else if (length(x) == 0) {
+    id <- integer()
+    n <- 0L
   } else {
     levels <- sort(unique(x), na.last = TRUE)
     id <- match(x, levels)